backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Crusazer <crusazer@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
cryptography = "^43.0.1"
pydantic = {extras = ["email"], version = "^2.9.1"}

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
